v1.19 1、修复一些已知问题，一些已知报错。 2、增加wavbuffer类，未完成

+++ YCore/LibSoundCard/ASIO/ASIOClient.h
@@ -39,7 +39,7 @@ public:
 	//获取采样率
 	int getSampleRate();
 	//设置采样率
-	TResult<void> setSampleRate(int sampleRate);
+	TResult<void> setSampleRate(long sampleRate);
 
 
 public:
